clang tidy (: (#10)
adds instructions for using clang tidy with this project (: along with clang-tidy files for the project and tests and a few initial nolint suppressions. the test specific clang tidy file was added to deal with the extra suppressions needed to deal with gtest macros. 

checks are pretty tentative right now. I basically just picked everything on this page that didnt seem to be specific to a particular organization/project https://clang.llvm.org/extra/clang-tidy/#id2. and then a few opinionated suppressions based on an initial output, so those are also pretty tentative. 

the casing options for the `readability-identifier-naming` are less tentative :p the main motivation for this casing is that different things should look different, also combined with my personal preference (: 

also snuck in a few misc fixes. we werent actually calling find package for gest in the test target 😅  also for whatever reason the conan install for gcc fails unless you disable x265. so both of those are fixed now (: