new action to run periodically on main (:  (#17)
this is basically the same as the current PR pipeline just with a clang tidy run on the whole project instead of clang tidy diff and clang format diff. the intent is to run once a week on the latest main. 

eventually this would probably run on every main commit but for now id rather save minutes (: 

also a few fixes/suppressions for clang tidy conformance (: